refactor: Moved models to Models folder
