+++ iOS-FakeNFT-Extended/UI/Stats/StatsModel.swift
@@ -7,16 +7,6 @@
 
 import Foundation
 
-struct UserStatsJsonModel: Sendable, Decodable, Identifiable {
-    let name: String?
-    let avatar: String?
-    let description: String?
-    let website: String?
-    let nfts: [String]?
-    let rating: String?
-    let id: String?
-}
-
 enum APIError: Error {
     case invalidURL
     case responseConvertationError
